week4/perception-tools/requirements.txt
# Core MCP dependencies
mcp>=0.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# HTTP and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.0.0

# Video processing
opencv-python>=4.8.0

# Data sources
wikipedia>=1.4.0
arxiv>=2.0.0

# Optional: Google Calendar integration
# google-auth-oauthlib>=1.0.0
# google-auth-httplib2>=0.1.0
# google-api-python-client>=2.0.0

# Optional: Notion integration
# notion-client>=2.0.0
